requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12.0,<4.0
mysqlclient>=2.0,<3.0
django-cors-headers>=3.5,<4.0
pandas>=1.3,<2.0